Create README.md
This is a short script that automates and installs the Heroes of Might and Magic V mod for simultaneous turns (http://heroescommunity.com/viewthread.php3?TID=39792).
Designed to be used for the less motivated of my friends so that they can actually install the mod.

It also downloads and opens the installer for Hamachi (https://www.vpn.net/) to create a virtual network so that online play is possible. (This functionality was discontinued for HOMMV because of it's age)

It does require some fairly specific files to be packaged with it in order for it to work. You would require the modded files:
1. H5-Game.NCF_pest.exe (the exe for the simultaneous turns mod)
2. NCF__Core.pak (This is a modded file from the "New Creatures Framework" mod) (http://www.heroesofmightandmagic.com/heroes5/modding_wiki/new_creatures_framework)
3. A HOMMV shortcut that gets placed on the desktop (just for ease of access)